Major change : - Node is now an interface : preparing for adding RemoteNode types - Old Node is now a LocalNode
